# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop API",
        "description": "Documentation for v1 api of pet store , this API is a recruitement test from Buckhill.",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/brands": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "View brands",
                "operationId": "36e96e64340e79d511d0af9c7965971a",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "uuid": {
                                            "type": "string",
                                            "example": "d166d772-7c61-4e53-95f1-8e0a27748e3a"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Brand title"
                                        },
                                        "slug": {
                                            "type": "string",
                                            "example": "Brand slug"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-07-09T03:11:38.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-07-09T03:11:38.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/brand/create": {
            "post": {
                "tags": [
                    "Brands"
                ],
                "summary": "store brand",
                "operationId": "06bedd60a7866202a202160764d25247",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Brand's title",
                                        "type": "string",
                                        "example": "Brand title example"
                                    },
                                    "slug": {
                                        "description": "Brand's slug",
                                        "type": "string",
                                        "example": "Brand slug example"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "brand stored with success"
                                        },
                                        "brand": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "uuid": {
                                                    "type": "string",
                                                    "example": "d166d772-7c61-4e53-95f1-8e0a27748e3a"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Brand title"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "Brand slug"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-07-09T03:11:38.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-07-09T03:11:38.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: 'Bearer {token}'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}